master - redis caching service for plans

+++ plans-service/services/plans-service.js
@@ -1,8 +1,24 @@
 const Plan = require('../models')['Plan']
+const CachingService = require('./caching-service')
 
 module.exports = class PlansService {
+  constructor() {
+    this.cachingService = new CachingService({
+      host: process.env.REDIS_HOST,
+      port: process.env.REDIS_PORT,
+      password: process.env.REDIS_PASSWORD,
+    })
+  }
+  
   async findAll(userId) {
-    return await Plan.findAll({where: {userId}})
+    let plans = await this.cachingService.getPlans(userId)
+    if (plans) {
+      return plans
+    } else {
+      let plans = await Plan.findAll({where: {userId}})
+      await this.cachingService.storePlans(plans, userId)
+      return plans
+    }
   }
   
   async findOne(id) {
@@ -10,10 +26,17 @@ module.exports = class PlansService {
   }
   
   async create(plan) {
+    await this.cachingService.purgeCache(plan.userId)
     return Plan.create(plan)
   }
   
   async deleteOne(id) {
-    return await Plan.destroy({where: {id}})
+    let plan = await this.findOne(id)
+    if (plan){
+      await this.cachingService.purgeCache(plan.userId)
+      return await Plan.destroy({where: {id}})
+    }else{
+      return Promise.resolve()
+    }
   }
 }
